Make VoieHeading component
